Add plotting library to the system
This plotting library provides some drawing capabilities on top of
gnuplot. At its current state it is very simplistic and so far only
acts as a wrapper around a gnuplot subprocess. Hence, there is not much
abstraction provided, but in turn also provides the full potential of
gnuplot to the program.