Add support to the plot_data method to plot multiple data sets in one figure
The data variable passed to the plot_data method can now either be a
list of lists if one only wants to plot one data set or a dict of list
of lists if one wants to plot multiple data sets. The plot_data method
will transfer all of them to gnuplot and then plot them in one figure.

The data sets are named either 'data' if only one is given or the keys
in the dict are used as names.